rag/requirements.txt
boto3==1.35.12
langchain==0.2.16
langchain-pinecone==0.1.3
pinecone-notebooks==0.1.1
langchain-aws==0.1.17
python-dotenv==1.0.1
